Bump puma to 4.3.5 due to security vulnerability
GHSA-x7jg-6pwg-fx5h
high severity
Vulnerable versions: >= 4.0.0, < 4.3.4
Patched version: 4.3.4
Impact
By using an invalid transfer-encoding header, an attacker could smuggle an HTTP response.

GHSA-w64w-qqph-5gxm
moderate severity
Vulnerable versions: >= 4.0.0, < 4.3.5
Patched version: 4.3.5
Impact
This is a similar but different vulnerability to the one patched in 3.12.5 and 4.3.4.

A client could smuggle a request through a proxy, causing the proxy to send a response back to another unknown client.

If the proxy uses persistent connections and the client adds another request in via HTTP pipelining, the proxy may mistake it as the first request's body. Puma, however, would see it as two requests, and when processing the second request, send back a response that the proxy does not expect. If the proxy has reused the persistent connection to Puma to send another request for a different client, the second response from the first client will be sent to the second client.